# spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Puja Booking Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/refunds/override": {
      "post": {
        "tags": [
          "admin-refunds"
        ],
        "summary": "Refund Override",
        "description": "Insert refunds row reason=admin_override \u2014 worker executes Razorpay (A-REFUND).",
        "operationId": "refund_override_v1_admin_refunds_override_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundOverrideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/admin/refunds": {
      "get": {
        "tags": [
          "admin-refunds"
        ],
        "summary": "List Refunds",
        "description": "List refunds by status \u2014 e.g. failed_permanent queue for ops.",
        "operationId": "list_refunds_v1_admin_refunds_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Refund status filter",
              "title": "Status"
            },
            "description": "Refund status filter"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRefundListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/promos": {
      "get": {
        "tags": [
          "admin-promos"
        ],
        "summary": "List Promos",
        "operationId": "list_promos_v1_admin_promos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPromoListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-promos"
        ],
        "summary": "Create Promo",
        "operationId": "create_promo_v1_admin_promos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPromoCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPromoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/promos/{promo_id}": {
      "put": {
        "tags": [
          "admin-promos"
        ],
        "summary": "Update Promo",
        "operationId": "update_promo_v1_admin_promos__promo_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "promo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Promo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPromoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPromoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminPromoCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 30,
            "minLength": 2,
            "title": "Code"
          },
          "discount_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Discount Pct"
          },
          "max_uses_per_user": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Uses Per User",
            "default": 1
          },
          "valid_from": {
            "type": "string",
            "format": "date-time",
            "title": "Valid From"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Until"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "discount_pct",
          "valid_from",
          "valid_until"
        ],
        "title": "AdminPromoCreate"
      },
      "AdminPromoListResponse": {
        "properties": {
          "promos": {
            "items": {
              "$ref": "#/components/schemas/AdminPromoOut"
            },
            "type": "array",
            "title": "Promos"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "promos"
        ],
        "title": "AdminPromoListResponse"
      },
      "AdminPromoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "discount_pct": {
            "type": "integer",
            "title": "Discount Pct"
          },
          "max_uses_per_user": {
            "type": "integer",
            "title": "Max Uses Per User"
          },
          "valid_from": {
            "type": "string",
            "format": "date-time",
            "title": "Valid From"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Until"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "redemption_count": {
            "type": "integer",
            "title": "Redemption Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "discount_pct",
          "max_uses_per_user",
          "valid_from",
          "valid_until",
          "is_active"
        ],
        "title": "AdminPromoOut"
      },
      "AdminPromoUpdate": {
        "properties": {
          "discount_pct": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Pct"
          },
          "max_uses_per_user": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses Per User"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "AdminPromoUpdate"
      },
...
      "AdminRefundListResponse": {
        "properties": {
          "refunds": {
            "items": {
              "$ref": "#/components/schemas/AdminRefundQueueItem"
            },
            "type": "array",
            "title": "Refunds"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "refunds"
        ],
        "title": "AdminRefundListResponse"
      },
      "AdminRefundQueueItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "booking_id": {
            "type": "string",
            "format": "uuid",
            "title": "Booking Id"
          },
          "payment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Payment Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "customer_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "booking_id",
          "payment_id",
          "amount",
          "status",
          "reason",
          "last_error",
          "attempt_count",
          "created_at",
          "customer_phone",
          "customer_name"
        ],
        "title": "AdminRefundQueueItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefundOverrideRequest": {
        "properties": {
          "payment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          },
          "booking_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "change_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Reason"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "RefundOverrideRequest"
      },
      "RefundOverrideResponse": {
        "properties": {
          "refund_id": {
            "type": "string",
            "format": "uuid",
            "title": "Refund Id"
          },
          "payment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Payment Id"
          },
          "booking_id": {
            "type": "string",
            "format": "uuid",
            "title": "Booking Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "refund_id",
          "payment_id",
          "booking_id",
          "amount",
          "status"
        ],
        "title": "RefundOverrideResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}